feat: sys_uart gpio template

+++ src/components/gpio/include/gpio_type.h
@@ -36,6 +36,6 @@ typedef enum{
 } gpio_level_t;
 
 typedef struct {
-    uint64_t pin_bit_mask; // 引脚位掩码，用于设置或读取多个引脚
+    uint32_t pin_bit_mask; // 引脚位掩码，用于设置或读取多个引脚
     gpio_mode_t mode; // GPIO模式，输入或输出
 } gpio_config_t;
